image_optimization_for_websites/requirements.txt
Pillow
pillow-avif-plugin